refactor: 보유 쿠폰 영역 & 상품 리스트 영역 컴포넌트로 분리
